Added more wrappers
