+++ src/main/java/ca/team2706/frc/robot/Robot.java
@@ -23,11 +23,9 @@ public class Robot extends TimedRobot {
     public void robotInit() {
         onStateChange(RobotState.ROBOT_INIT);
 
-        oi = new OI();
+        OI.init();
 
-        // Usb Camera
-        // UsbCamera camera = CameraServer.getInstance().startAutomaticCapture(0);
-        // Runtime.getRuntime().addShutdownHook(new Thread(camera::free));
+        
     }
 
     /**
